+++ libs/ymir-core/src/ymir/media/loader/loader_bin_cue.cpp
@@ -332,38 +332,15 @@ bool Load(std::filesystem::path cuePath, Disc &disc, bool preloadToRAM) {
                     prevTrack.indices.back().endFrameAddress = track.startFrameAddress + frameAddress - 1;
                 }
 
-                // If the pregap in audio tracks is actually silent, skip it
-                if (hasPregap && track.controlADR == 0x01) {
+                // Skip pregap in audio tracks or data tracks not starting at the beginning of a file
+                if (hasPregap) {
                     const uintmax_t pregapEnd = TimestampToFrameAddress(m, s, f);
                     const uintmax_t pregapStart = hasIndex0
                                                       ? TimestampToFrameAddress(prevM, prevS, prevF)
                                                       : pregapEnd - TimestampToFrameAddress(pregapM, pregapS, pregapF);
-                    pregapLength = pregapEnd - pregapStart;
-                    std::vector<uint8> sector{};
-                    sector.resize(track.sectorSize);
-                    bool isPregapSilent = true;
-                    /*for (uintmax_t frame = 0; frame < pregapLength; ++frame) {
-                        binaryReader->Read(binFileOffset + frame * track.sectorSize, track.sectorSize, sector);
-                        // Some dumps contain an extra data sector at the start of the first audio track.
-                        // If found, skip it
-                        size_t pregapCheckOffset = 0;
-                        static constexpr uint8 kSyncHeader[] = {0x00, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-                                                                0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x00};
-
-                        if (std::equal(sector.begin(), sector.begin() + std::size(kSyncHeader),
-                                       std::begin(kSyncHeader))) {
-                            pregapCheckOffset += 2352;
-                        }
-                        if (std::any_of(sector.begin() + pregapCheckOffset, sector.end(),
-                                        [](uint8 v) { return v != 0; })) {
-                            isPregapSilent = false;
-                            break;
-                        }
-                    }*/
-                    if (isPregapSilent) {
-                        const uintmax_t delta = pregapEnd - pregapStart;
-                        // fmt::println("BIN/CUE: Track {} has silent pregap of {} frames; skipping", currTrackIndex,
-                        //              delta);
+                    const uintmax_t delta = pregapEnd - pregapStart;
+                    if (track.controlADR == 0x01 || binFileOffset > 0) {
+                        // fmt::println("BIN/CUE: Track {} has pregap of {} frames; skipping", currTrackIndex, delta);
                         binFileOffset += delta * track.sectorSize;
                     }
                 }